requirements.txt
accelerate==1.11.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.4.0
audiomentations==0.43.1
audioread==3.1.0
babel==2.17.0
beautifulsoup4==4.14.2
# bitsandbytes==0.48.2
black==25.9.0
bleach==6.3.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
datasets==4.4.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
einops==0.8.1
executing==2.2.1
fastjsonschema==2.21.2
ffmpeg-python==0.2.0
filelock==3.20.0
fonttools==4.60.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.10.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
graphviz==0.21
h11==0.16.0
hf-xet==1.2.0
hf_transfer==0.1.9
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
imageio-ffmpeg==0.6.0
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
isort==7.0.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
lark==1.3.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.45.1
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mistune==3.1.4
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.18
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.7
notebook_shim==0.2.4
numba==0.62.1
numpy==2.3.5
numpy-minmax==0.5.0
numpy-rms==0.6.0
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pathspec==0.12.1
peft==0.18.0
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
pooch==1.8.2
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
pytest==8.4.2
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
python-stretch==0.3.1
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.28.0
ruff==0.14.3
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==5.1.2
sentencepiece==0.2.1
sentry-sdk==2.45.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8
soxr==0.5.0.post1
stack-data==0.6.3
sympy==1.14.0
terminado==0.18.1
threadpoolctl==3.6.0
timm==1.0.21
tinycss2==1.4.0
tokenizers==0.22.1
torch==2.9.0
torchaudio==2.9.0
torchcodec==0.8.1
torchview==0.2.7
torchvision==0.24.0
torchviz==0.0.3
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
triton==3.5.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uv==0.9.7
wandb==0.23.0
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
wheel==0.45.1
widgetsnbextension==4.0.14
xxhash==3.6.0
yarl==1.22.0
yt-dlp==2025.11.12
